refactor(setup_presets): extract ReolinkClient into reolink.py module
